linux: tweak idle energy handling
For proper handling of idle states we need the Crust firmware
for the SCP, which is not ready for mass usage yet.

Therefore, add bits of support I know that will be useful
for when that time comes.

Also, experimental observations have shown that adding
a trip point at 75ºC to barely slow down the CPU clock
is useless.