pinetainer: readd, this time with Xen
PineTainer is still preliminary, but the Xen hypervisor
works and boots the Linux kernel as usual. I've managed
to get it booting with 1 GiB of RAM for dom0. With 512 MiB
it didn't work (the kernel hangs early at "initmem setup
node 0"), but that might be due to a bad boot script.

I will improve this soon.