pinetainer: use hostapd as RADIUS server instead
Tests have shown that the Realtek 8723BS chip is pretty bad when used as
an access point, as it doesn't support HT rates, ACS, country-specific
regulatory settings, transmission power and other desireable features.
The ISP gateway, in our case, has a Wi-Fi chip that is in all respects
better, and has a built-in RADIUS client that can be used for 802.11x
EAP. That way we use the Pine to provide extra security to the ISP
gateway Wi-Fi, and take advantage of its superior radio chip. In
particular, our Wi-Fi setup is not vulnerable to KRACK and offline PSK
cracking via PMKID this way.

This change also makes the second bridge unnecessary, as it was only
used to host WLAN guests. Also, when using hostapd for its RADIUS server
only, it can be run without root priviledges.