pinetainer: configure firewall, NAT, shaping, etc.
These changes should end up the Pine network configuration, but they
need to be tested on the actual Pine hardware with actual virtual
machines later, to be sure that they work as they are intended to, which
should be a relatively easy task.

We also make the chrony service start sooner, so that the Pine has a
correct date and time before offering services to the network, which may
be sensitive to this.

We also drop the knock package, as nftables provides means to integrate
the port knocking logic pretty well in the kernel, which has
performance, memory consumption and failure handling advantages.

Also, when bringing up eth0, we trust the MAC address we learned for the
gateway and make that permanent, instead of hardcoding a MAC address.
This should make a gateway change easier without much security concerns,
as our threat model does not include someone in the LAN doing ARP
spoofing just when the Pine is brought up.