Workaround nftables resolution of "-" to /dev/sdtin
nftables considers "-" as an alias for the symlink /dev/stdin, which is
created by udev or mdev. However, if devtmpfs is mounted on /dev and
udev or mdev are not running, the symlink is not created, and the
command fails.

As we run network scripts before mdev, use /proc/self/fd/0 directly
instead, which is the target of the symlink created by mdev, and while
we are at it clean up the post-build scripts of related things.