pinetainer: configure OpenSSH server
This configuration allows remote logins via SSH to a new
"admin" account, which is by default disabled. The authentication
requires both a public key and a password, which is meant to
provide a kind of 2FA auth (the key that the user has, and a
password that the user knows).

This is yet to be tested, and the sudoers file must be edited
to allow this admin user to become root for administrative tasks.